Using `dense_e` metric option.  This *will not work* under default Stan.
By default, Stan estimates the posterior scale independently in each dimension,
and the draws momenta according to the appropriate scales.  This is sometimes
called using a "diagonal" mass matrix.  It does not work well if there are
correlations among the variables.  Stan has included an option for estimating a
full posterior covariance and drawing momenta according to the corresponding
"dense" mass matrix.  This can account for the correlations among variables,
but becomes very computationally expensive in high dimensions (inverting the
covariance scales as O(Ndim^3)!).

Unfortunately, there is a bug in the way that Stan inverts the estimated
covariance matrix that makes sampling with a dense mass matrix completely
unstable (one wonders how testing didn't catch this, since it's *totally
obvious* once you try even simple correlated problems).  The commit
ba5638b82a85df71847484f542b9d55d2e067b29 in the *stan* (not PyStan!) repo fixes
this bug, but isn't yet on any mainline branch.  You will need to cherry pick
this commit into the stan sub-module and re-build pystan if you want this
sampling to work:

    cd [pystan-directory]
    cd pystan/stan
    git cherry-pick ba5638b82a85df71847484f542b9d55d2e067b29
    cd ../..
    rm -rf build # Clean, just in case
    python setup.py install

It's worth it---even the weak correlations in this problem can lead to order of
magnitude slowdown under PyMC3, so expect similar improvements in performance
with Stan.  And, of course, compling the 5-year analysis under PyMC3 is
completely impossible!